Squashed commit of the following:
commit efa7750ba35e3e32ba005a01cf53aab949c6993c
Author: AJ Slater <aj@slater.net>
Date:   Mon Oct 5 12:27:34 2020 -0700

    componentize browser-main

commit 3e9caadb2ffa7e7b679ab813d4c5acb4a7f366cb
Author: AJ Slater <aj@slater.net>
Date:   Mon Oct 5 12:06:01 2020 -0700

    oops forgot to commit settings file for module change

commit ca5168afd78bf05bdc7572120c746beb2e93ef22
Author: AJ Slater <aj@slater.net>
Date:   Mon Oct 5 12:05:31 2020 -0700

    settings module

commit 54af69ae2562fdb7b101ef57d5ebbf9dd8088482
Author: AJ Slater <aj@slater.net>
Date:   Sun Oct 4 20:25:01 2020 -0700

    componentize browser
