+++ codex/codex_settings/README.md
@@ -1,0 +1,3 @@
+# codex_settings
+
+I moved the functions that service the django settings file into their own module just to clean up the readability of the settings file.
